Adding a way to provide a backend specific create table statement for the migration table
Different databases uses different sql dialects. While the current
table statement for the migration table works on all supported
backends it is a show stopper while implementing a new backend for
supporting a oracle database.

Technically this commit is a breaking change. This may affect only
custom backend implementation. We are not aware of any such
implementation.